requirements.txt
pandas
requests
confluent-kafka
pyspark
SQLAlchemy
python-dotenv
psycopg2-binary